+++ app/src/pages/settings/datasets/DatasetLabelsSettingsCard.tsx
@@ -1,0 +1,28 @@
+import { graphql, useFragment } from "react-relay";
+
+import { Card } from "@phoenix/components";
+import { NewDatasetLabelButton } from "@phoenix/components/dataset/NewDatasetLabelButton";
+
+import { DatasetLabelsSettingsCardFragment$key } from "./__generated__/DatasetLabelsSettingsCardFragment.graphql";
+import { DatasetLabelsTable } from "./DatasetLabelsTable";
+
+export function DatasetLabelsSettingsCard({
+  query,
+}: {
+  query: DatasetLabelsSettingsCardFragment$key;
+}) {
+  const data = useFragment<DatasetLabelsSettingsCardFragment$key>(
+    graphql`
+      fragment DatasetLabelsSettingsCardFragment on Query {
+        ...DatasetLabelsTableFragment
+      }
+    `,
+    query
+  );
+
+  return (
+    <Card title="Dataset Labels" extra={<NewDatasetLabelButton />}>
+      <DatasetLabelsTable query={data} />
+    </Card>
+  );
+}